backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.2
django-cors-headers==4.4.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0
